Extract haul data

Extract catch data for cpue by seabed area biomass

extract_catch_data_cpue_est(catch_data, release_data, catch_season = NULL,
  measure, mean_fish_weight = NULL, target_species)

Arguments

catch_data
catch data at haul-by-haul resolution that includes a data frame with columns of data in the following order: Unique Catch ID record reference, Fishing Season,cruise id, set id, species code, retained catch quantity (kg),long line length

release_data
release data at the haul-by-haul resolution that includes a data frame with columns of data in the following order: Fishing Season,cruise id, set id, species code and fish weights if available and measure="weights", if fish weights are not available and measure="counts" these weights are estimated

catch_season
is the CCAMLR fishing seasons of interest where the default = NULL uses all the catch data seasons included in the catch_data dataframe

measure
"counts" if only fish counts are available or "weights" if fish weights are required

mean_fish_weight
is the mean weight of a fish to estimate the fish release weights per haul if measure= "counts" otherwise the default=NULL

target_species
is the species code for the species that is targetted in the research block or reference area