Simulated tagged fish recapture data from hypothetical Research blocks for vignette examples

A dataset that aims to simulate the tagged fish recapture data necessary for toothfish local biomass estimates that is stored in CCAMLR's cdb database

recapture_data_sim_RB

Format

A dataframe with 12 rows and 7 columns

SEASON_RELEASE

The simulated CCAMLR fishing season (i.e. December from the previous calendar year through to the end of November from the current calendar year) the tagged fish was released

SEASON_RECAPTURE

The simulated CCAMLR fishing season (i.e. December from the previous calendar year through to the end of November from the current calendar year) the tagged fish was recaptured

CRUISE_ID_RECAPTURE

An ID used by Observers that is recorded with tagged fish recaptures

SET_ID_RECAPTURE

An ID used by Observers that is recorded with recaptured fish for each set-haul

SPECIES_CODE_RECAPTURE

The three letter code used by CCAMLR to record the species of a tagged fish recapture. In the simulated data example this represent either Antarctic toothfish = "TOA" or Patagonian toothfish = "TOP"

RESEARCH_BLOCK_CODE_RECAPTURE

A simulated hypothetical Research block code that the tagged fish recapture data fall within

RESEARCH_BLOCK_CODE_RELEASE

A simulated hypothetical Research block code that the tagged fish release data fall within