Single tag release estimate of population size

Estimate population size from single release tag data

single_release(tags, catch, recaps, mean_wt = 0, prior_recaps = 0, method,
  unit, type, tag_mort = 0, reporting = 1, nat_mort = 0,
  chronic_shed = 0, chronic_mort = 0)

Arguments

tags
number of marked animals released (note these must be whole numbers for the bootstrap to work)

catch
vector of the number or weight of animals captured and checked for tags by haul on the second survey

recaps
vector of number of marked animals recaptured by haul

mean_wt
mean weight of a single fish (optional argument for Chapman weight method)

prior_recaps
vector of subsequent recaptures (first year this represents within season recaptures)

method
method

unit
measurement unit (numbers, kg, tonnes)

type
do we assume fishing occurs over a short period (Ricker type 1) or is extended over the year and competes with natural mortality and tag sheddding (Ricker type 2)

tag_mort
initial tag-induced mortality

reporting
vector of tag reporting rate

nat_mort
vector of natural mortality (instantaneous)

chronic_shed
vector of chronic (ongoing) tag shedding

chronic_mort
vector of chronic (ongoing) tag-induced mortality

Details

For a Type 1 fishery the adjusted releases are calculate as follows

1) Initial tag-induced mortality is applied

2) Within season recaptures divided by the tag reporting rate are removed

3) Natural mortality and chronic tag loss and tag-induced mortality is applied.

For a fishery that operates throughout the year (Ricker Type 2) fishing and natural mortality competed to deplete the tagged population. The adjusted releases are calculated as follows

1) Initial tag-induced mortality is applied

2) Half of the within season recaptures divided by the tag reporting rate are removed

3) Half of the natural mortality and chronic tag loss and tag-induced mortality is applied

4) The remaining within season recaptures divided by the tag reporting rate are removed

5) The remaining natural mortality and chronic tag loss and tag-induced mortality is applied.

Examples

## Chapman estimates of population numbers
single_release(tags=100, catch=200, recaps=5, method="Chapman",
               type=1, unit="numbers")

#> Chapman estimate of population numbers 
#>     N_hat     var_N 
#>    3382.5 1492365.2 

single_release(tags=100, catch=200, recaps=5, method="Chapman",
               type=1, unit="numbers", tag_mort=0.1)

#> Chapman estimate of population numbers 
#>       N_hat       var_N 
#>    3063.705 1216347.233 

single_release(tags=100, catch=200, recaps=5, method="Chapman", type=1,
               unit="numbers", tag_mort=0.1, reporting=0.8)

#> Chapman estimate of population numbers 
#>      N_hat      var_N 
#>   2535.308 692050.552